Refactor macros out of Config
