+++ lib/bootleg/task.ex
@@ -22,13 +22,13 @@ defmodule Bootleg.Task do
 
   Technically speaking, any module in the namespace `Bootleg.Tasks` that exports a `load/0` function will
   be discovered and executed by Bootleg automatically. This usage is not recommended unless you need to
-  do work before `use Bootleg.Config`.
+  do work before `use Bootleg.DSL`.
 
   ```
   defmodule Bootleg.Tasks.Other do
     use Bootleg.Task
     def load do
-      use Bootleg.Config
+      use Bootleg.DSL
 
       task :other do
         IO.puts "World?"
