+++ lib/bootleg/tasks/restart.exs
@@ -1,5 +1,5 @@
 alias Bootleg.{UI, Config}
-use Bootleg.Config
+use Bootleg.DSL
 
 task :restart do
   remote :app do
