+++ lib/bootleg/tasks/stop.exs
@@ -1,5 +1,5 @@
 alias Bootleg.{UI, Config}
-use Bootleg.Config
+use Bootleg.DSL
 
 task :stop do
   app_name = Config.app()
