+++ test/fixtures/build_me/config/deploy.exs
@@ -1,1 +1,1 @@
-use Bootleg.Config
+use Bootleg.DSL
