+++ test/fixtures/build_me/config/deploy/production.exs
@@ -1,1 +1,1 @@
-use Bootleg.Config
+use Bootleg.DSL
